refactor: Event 기반 코드 listener로 리팩토링
