+++ src/main/java/com/hanium/mom4u/external/s3/FileStorageService.java
@@ -1,17 +1,9 @@
 package com.hanium.mom4u.external.s3;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hanium.mom4u.domain.news.repository.NewsJdbcRepository;
-import com.hanium.mom4u.domain.news.service.S3JsonImportEvent;
-import com.hanium.mom4u.global.crawling.dto.CrawlingResultDto;
-import com.hanium.mom4u.global.exception.GeneralException;
-import com.hanium.mom4u.global.response.StatusCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Service;
 import software.amazon.awssdk.auth.credentials.*;
-import software.amazon.awssdk.core.ResponseInputStream;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.*;
@@ -19,10 +11,8 @@ import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 import software.amazon.awssdk.services.s3.presigner.model.PutObjectPresignRequest;
 
 
-import java.io.IOException;
 import java.net.URL;
 import java.time.Duration;
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -40,11 +30,6 @@ public class FileStorageService {
     @Value("${spring.cloud.aws.s3.secret-key}") // optional - 로컬에서만 필요
     private String secretKey;
 
-
-    private final S3Client s3Client;
-    private final ObjectMapper objectMapper;
-    private final NewsJdbcRepository newsJdbcRepository;
-    private static final String NEWS_KEY = "data/preprocessed/";
 
     public String generatePresignedPutUrl(String objectKey) {
         S3Presigner presigner;
@@ -95,39 +80,5 @@ public class FileStorageService {
 
         s3Client.deleteObject(deleteRequest);
         s3Client.close();
-    }
-
-    @EventListener
-    public void importJsonFromS3(S3JsonImportEvent event) {
-
-        ListObjectsV2Request listRequest = ListObjectsV2Request.builder()
-                .bucket(event.getBucketName())
-                .prefix(NEWS_KEY)
-                .build();
-
-        ListObjectsV2Response listResponse = s3Client.listObjectsV2(listRequest);
-
-        List<S3Object> objectList = listResponse.contents();
-
-        for (S3Object s3Object : objectList) {
-            String key = s3Object.key();
-
-            if (!key.endsWith(".json")) {
-                continue; // skip
-            }
-            ResponseInputStream<GetObjectResponse> s3InputStream = s3Client.getObject(GetObjectRequest.builder()
-                    .bucket(bucketName)
-                    .key(key)
-                    .build());
-
-            try {
-                CrawlingResultDto dto = objectMapper.readValue(s3InputStream, CrawlingResultDto.class);
-
-                newsJdbcRepository.save(dto);
-                System.out.printf("{} 저장 성공", dto.getPostId());
-            } catch (IOException e) {
-                throw new GeneralException(StatusCode.JSON_PARSING_ERROR);
-            }
-        }
     }
 }
